refactor: remove `client` sub-module
