+++ crates/rust-client/src/mock.rs
@@ -28,33 +28,31 @@ use tonic::{Response, Status};
 use uuid::Uuid;
 
 use crate::{
-    client::{
-        rpc::{
-            AccountDetails, NodeRpcClient, NodeRpcClientEndpoint, NoteDetails,
-            NoteInclusionDetails, StateSyncInfo,
-        },
-        store_authenticator::StoreAuthenticator,
-        sync::FILTER_ID_SHIFT,
-        transactions::{
-            prepare_word,
-            transaction_request::{PaymentTransactionData, TransactionTemplate},
-        },
-        Client,
-    },
     config::RpcConfig,
     errors::RpcError,
-    rpc::generated::{
-        self,
-        account::AccountId as ProtoAccountId,
-        block_header::BlockHeader as NodeBlockHeader,
-        note::NoteSyncRecord,
-        requests::SyncStateRequest,
-        responses::{NullifierUpdate, SyncStateResponse},
+    rpc::{
+        generated::{
+            self,
+            account::AccountId as ProtoAccountId,
+            block_header::BlockHeader as NodeBlockHeader,
+            note::NoteSyncRecord,
+            requests::SyncStateRequest,
+            responses::{NullifierUpdate, SyncStateResponse},
+        },
+        AccountDetails, NodeRpcClient, NodeRpcClientEndpoint, NoteDetails, NoteInclusionDetails,
+        StateSyncInfo,
     },
     store::{
         sqlite_store::{config::SqliteStoreConfig, SqliteStore},
         InputNoteRecord,
     },
+    store_authenticator::StoreAuthenticator,
+    sync::FILTER_ID_SHIFT,
+    transactions::{
+        prepare_word,
+        transaction_request::{PaymentTransactionData, TransactionTemplate},
+    },
+    Client,
 };
 
 pub type MockClient =
