+++ crates/rust-client/src/note_screener.rs
