+++ crates/rust-client/src/rpc/tonic_client/domain/blocks.rs
