+++ crates/rust-client/src/rpc/tonic_client/domain/digest.rs
