+++ crates/rust-client/src/rpc/tonic_client/domain/merkle.rs
