+++ crates/rust-client/src/store/mod.rs
@@ -10,11 +10,9 @@ use miden_objects::{
 use winter_maybe_async::{maybe_async, maybe_await};
 
 use crate::{
-    client::{
-        sync::StateSyncUpdate,
-        transactions::{TransactionRecord, TransactionResult},
-    },
     errors::StoreError,
+    sync::StateSyncUpdate,
+    transactions::{TransactionRecord, TransactionResult},
 };
 
 pub mod data_store;
