+++ crates/rust-client/src/store_authenticator.rs
