+++ crates/rust-client/src/transactions/asm/transaction_scripts/auth_consume_notes.masm
