+++ crates/rust-client/src/transactions/asm/transaction_scripts/auth_send_asset.masm
