+++ crates/rust-client/src/transactions/asm/transaction_scripts/distribute_fungible_asset.masm
